backend/pyproject.toml
[project]
name = "job-applier"
version = "0.1.0"
description = "CV and cover letter optimizer for job applications"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "anthropic>=0.49",
    "openai>=1.65",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "weasyprint>=63",
    "httpx>=0.28",
    "beautifulsoup4>=4.13",
    "python-multipart>=0.0.20",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-ulid>=3.0",
    "jinja2>=3.1",
    "pillow>=11",
    "websockets>=14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_applier"]
